refactor: remove duplicate `NodeType` struct
This struct had been duplicated, so one of them was deleted and the other was moved to sit alongside
the other enums in the main `lib` module.
